+++ src/mefit/Firmware/Helper.cs
@@ -4,9 +4,7 @@
 // Helper.cs
 // Released under the GNU GLP v3.0
 
-using Mac_EFI_Toolkit.Tools;
 using System;
-using System.Linq;
 using System.Runtime.InteropServices;
 
 namespace Mac_EFI_Toolkit.Firmware