+++ src/mefit/Logger.cs
@@ -8,7 +8,6 @@ using Mac_EFI_Toolkit.UI;
 using System;
 using System.Diagnostics;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Windows.Forms;
 
 namespace Mac_EFI_Toolkit